Clean up formatting
- Space after `-`/`+` in method declarations
- Space after `if`
- Consistent usage of pointer notation (pointer goes with variable)
- Newlines at end of files
- Use conventional init pattern (init self before `if (self)`)
- Remove one empty init

+++ Barmoji.h
@@ -9,35 +9,35 @@
 #define kPrefDomain "com.cpdigitaldarkroom.barmoji"
 
 @interface EMFEmojiToken : NSObject
-@property (nonatomic,copy) NSString * string;
+@property (nonatomic, copy) NSString *string;
 @end
 
 @interface EMFEmojiPreferences : NSObject
--(NSArray *)allRecents;
--(NSArray *)recentEmojis;
+- (NSArray *)allRecents;
+- (NSArray *)recentEmojis;
 @end
 
 @class BarmojiCollectionView;
 
 @interface TUIPredictionView : UIView // iOS 13 +
-@property (nonatomic,retain) BarmojiCollectionView* barmoji;
+@property (nonatomic, retain) BarmojiCollectionView *barmoji;
 - (void)toggleBarmoji;
 @end
 
 @interface UIKeyboardImpl : UIView
 
-+(id)activeInstance;
--(void)insertText:(id)arg1 ;
++ (id)activeInstance;
+- (void)insertText:(id)arg1;
 @end
 
 @interface UIKeyboardEmoji
-@property (nonatomic,retain) NSString* emojiString;
-+(id)emojiWithString:(id)arg1 withVariantMask:(unsigned long long)arg2 ;
+@property (nonatomic, retain) NSString *emojiString;
++ (id)emojiWithString:(id)arg1 withVariantMask:(NSUInteger)arg2;
 @end
 
 @interface UIKeyboardEmojiCollectionViewCell : UICollectionViewCell
-@property (nonatomic,copy) UIKeyboardEmoji* emoji;
-@property (assign, nonatomic) long long emojiFontSize;
+@property (nonatomic, copy) UIKeyboardEmoji *emoji;
+@property (assign, nonatomic) NSInteger emojiFontSize;
 @end
 
 @interface UIKeyboardEmojiKeyDisplayController : NSObject
@@ -45,23 +45,23 @@
 @end
 
 @interface UIKeyboardPredictionView : UIView
-@property (nonatomic,retain) BarmojiCollectionView* barmoji;
+@property (nonatomic, retain) BarmojiCollectionView *barmoji;
 @end
 
 @interface UIKeyboardDockItemButton: UIButton
 @end
 
 @interface UIKeyboardDockItem : NSObject
-@property (nonatomic,retain) UIKeyboardDockItemButton * button; 
+@property (nonatomic, retain) UIKeyboardDockItemButton *button; 
 @end
 
 @interface UIKeyboardDockView : UIView
-@property (nonatomic,retain) BarmojiCollectionView* barmoji;
--(id)_keyboardLayoutView;
+@property (nonatomic, retain) BarmojiCollectionView *barmoji;
+- (id)_keyboardLayoutView;
 @end
 
 @interface UISystemKeyboardDockController : UIViewController
-@property (nonatomic,retain) UIKeyboardDockView* dockView;
+@property (nonatomic, retain) UIKeyboardDockView *dockView;
 @end
 
 @interface TIKeyboardCandidateSingle : NSObject
